requirements.txt
numpy==1.24.4
streamlit==1.36.0
ultralytics==8.3.3
pillow==9.4.0
fastapi==0.100.0
uvicorn==0.23.2
requests==2.31.0
python-multipart==0.0.6
tempfile==0.1.0